feat(session): headers
- Implemented Poll Host header to abstract toolbar
- Implemented QuestionBox for hosts to display question
- Implemented Poll Participate header to abstract toolber
- Removed verbose console debugs
